requirements.txt
streamlit==1.28.1
PyPDF2==3.0.1
python-docx==0.8.11
openai==1.3.7
langchain==0.0.350
langchain-openai==0.0.2
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.24.3
plotly==5.17.0
Pillow==10.1.0
fitz==0.0.1.dev2
pymupdf==1.23.8
docx2txt==0.8
textstat==0.7.3
spacy==3.7.2 